Facundo

+++ src/main/java/productos/API/Controllers/ClienteController.java
@@ -1,10 +1,16 @@
 package productos.API.Controllers;
 
 
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.annotation.Validated;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 import productos.API.Model.Entity.Cliente;
 import productos.API.Model.Entity.ClienteDTO;
@@ -12,6 +18,7 @@ import productos.API.Model.Payload.Response;
 import productos.API.Service.IClienteService;
 import productos.API.Service.Implementaciones.ClienteIMPL;
 
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
@@ -24,42 +31,51 @@ public class ClienteController {
     @Autowired
     private IClienteService clienteService;
 
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    public ResponseEntity<?> handleValidationExceptions(MethodArgumentNotValidException ex) {
+        String errorMessage = ex.getBindingResult().getAllErrors().get(0).getDefaultMessage();
+        return new  ResponseEntity<>(Response.builder().mensaje(errorMessage).object(null).build(), HttpStatus.BAD_REQUEST);
+    }
+
+
     @PostMapping("cliente")
-    public ResponseEntity<?> create(@RequestBody ClienteDTO clienteDto){
+    public ResponseEntity<?> create(@RequestBody @Valid ClienteDTO clienteDto){
         Cliente clienteCreate = null;
         try{
 
-            String nombreEnMayusculas = clienteDto.getNombre_Completo().toUpperCase();
-            clienteDto.setNombre_Completo(nombreEnMayusculas);
-             clienteCreate = clienteService.save(clienteDto);
-            ClienteDTO cliente = ClienteDTO.builder()
-                    .id(clienteCreate.getId())
-                    .Mail(clienteCreate.getMail())
-                    .Nombre_Completo(clienteCreate.getNombre_Completo())
-                    .Telefono(clienteCreate.getTelefono())
-                    .build();
 
-            return new ResponseEntity<>(
-                    Response.builder()
-                            .mensaje("creado correctamente")
-                            .object(cliente)
-                            .build()
-                    , HttpStatus.CREATED);
 
-        }catch (DataAccessException dtx){
+                clienteCreate = clienteService.save(clienteDto);
+                ClienteDTO cliente = ClienteDTO.builder()
+                        .id(clienteCreate.getId())
+                        .Mail(clienteCreate.getMail())
+                        .Nombre_Completo(clienteCreate.getNombre_Completo())
+                        .Telefono(clienteCreate.getTelefono())
+                        .Direccion(clienteCreate.getDireccion())
+                        .Dni(clienteCreate.getDni())
+                        .build();
+
+                return new ResponseEntity<>(
+                        Response.builder()
+                                .mensaje("Cliente creado correctamente!!!")
+                                .object(cliente)
+                                .build()
+                        , HttpStatus.CREATED);
+            }catch (Exception e){
 
             return  new ResponseEntity<>(
                     Response.builder().
-                            mensaje(dtx.getMessage()).
+                            mensaje("Error al crear el cliente").
                             object(null)
                             .build(),
-                    HttpStatus.METHOD_NOT_ALLOWED);
+                    HttpStatus.BAD_REQUEST);
         }
 
     }
 
     @PutMapping("cliente/{id}")
-    public ResponseEntity<?> update(@RequestBody ClienteDTO clienteDto,@PathVariable Integer id){
+    public ResponseEntity<?> update(@RequestBody @Valid ClienteDTO clienteDto,@PathVariable Integer id){
         Cliente clienteUpdate = null;
 
         try{
@@ -71,18 +87,20 @@ public class ClienteController {
                         .Mail(clienteUpdate.getMail())
                         .Nombre_Completo(clienteUpdate.getNombre_Completo())
                         .Telefono(clienteUpdate.getTelefono())
+                        .Direccion(clienteUpdate.getDireccion())
+                        .Dni(clienteUpdate.getDni())
                         .build();
 
                 return new ResponseEntity<>(
                         Response.builder()
-                                .mensaje("actualizado correctamente")
+                                .mensaje("Cliente actualizado correctamente!!!")
                                 .object(cliente)
                                 .build()
                         , HttpStatus.CREATED);
             }else{
                 return  new ResponseEntity<>(
                         Response.builder().
-                                mensaje("no se encontro nada gil").
+                                mensaje("No se encontro este cliente").
                                 object(null)
                                 .build(),
                         HttpStatus.NOT_FOUND);
@@ -91,29 +109,36 @@ public class ClienteController {
 
             return  new ResponseEntity<>(
                     Response.builder().
-                            mensaje(dtx.getMessage()).
+                            mensaje("Error al editar el cliente").
                             object(null)
                             .build(),
-                    HttpStatus.METHOD_NOT_ALLOWED);
+                    HttpStatus.BAD_REQUEST);
         }
 
     }
 
-    @DeleteMapping("cliente/{id}")
-    public ResponseEntity<?> delete(@PathVariable Integer id){
+    @DeleteMapping("cliente")
+    public ResponseEntity<?> delete(@RequestBody ArrayList<Integer> ids){
 
-
+        ArrayList<Cliente> clientes = new  ArrayList<Cliente>();
         try{
-            Cliente cliente = clienteService.findById(id);
-            clienteService.delete(cliente);
-            return new ResponseEntity<>(Response.builder().mensaje("borrado correctamente").object(null).build(), HttpStatus.NO_CONTENT);
+             ids.forEach(id -> {
+                     clientes.add(clienteService.findById(id));
+
+            });
+             clientes.forEach(cliente -> {
+                 if(cliente != null) {
+                     clienteService.delete(cliente);
+                 }
+             });
+            return new ResponseEntity<>(Response.builder().mensaje("Cliente borrado correctamente!!!").object(null).build(), HttpStatus.OK);
 
         }catch (DataAccessException dtx){
             return new ResponseEntity<>(
                     Response.builder()
-                    .mensaje(dtx.getMessage())
+                    .mensaje("No se pudo eliminar el cliente")
                     .object(null)
-                    .build(),HttpStatus.METHOD_NOT_ALLOWED);
+                    .build(),HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -138,7 +163,7 @@ public class ClienteController {
                     .build();
             return new ResponseEntity<>(
                     Response.builder()
-                            .mensaje("buscado correctamente")
+                            .mensaje("Cliente encontrado correctamente!!!")
                             .object(cliente)
                             .build()
                     , HttpStatus.OK);
@@ -146,15 +171,16 @@ public class ClienteController {
     }
     @GetMapping("clientes")
     public ResponseEntity<?> findAll() {
+
         Iterable<Cliente> clientes = clienteService.findAll();
         if (clientes != null && clientes.iterator().hasNext()) {
             return new ResponseEntity<>(Response.builder()
-                    .mensaje("")
+                    .mensaje("Clientes encontrados con exito!!!")
                     .object(clientes)
                     .build(), HttpStatus.OK);
         } else {
             return new ResponseEntity<>(Response.builder()
-                    .mensaje("No se encontró nada")
+                    .mensaje("No hay clientes existentes")
                     .object(null)
                     .build(), HttpStatus.NOT_FOUND);
         }
